chore: PR 5 — delete dead-code components + colocated tests + rename FigmaOverviewView
Implements PR 5 of the synthesis arc (codex/synthesis-dead-code-purge).

Deleted components (all confirmed orphan via grep — no import statements
reach them from any live tree):

  src/lib/components/OverviewView.svelte           (494 lines — dial-era orphan)
  src/lib/components/ChronographDial.svelte        (792)
  src/lib/components/CausalVerdictStrip.svelte     (675)
  src/lib/components/RacingStrip.svelte            (377)
  src/lib/components/RunStorylineCard.svelte       (970)
  src/lib/components/OverviewSubtabStrip.svelte    (109)
  src/lib/components/EndpointRail.svelte           (268)
  src/lib/components/EventFeed.svelte              (212)
  src/lib/components/LatencyLegend.svelte          ( 50)

Deleted colocated tests (each imported the component being deleted):

  tests/unit/components/ChronographDial.test.ts    (322 lines)
  tests/unit/components/CausalVerdictStrip.test.ts (310)
  tests/unit/components/RacingStrip.test.ts        (119)
  tests/unit/components/RunStorylineCard.test.ts   (272)
  tests/unit/components/EndpointRail.test.ts       (186)
  tests/unit/components/EventFeed.test.ts          ( 62)

Renamed:

  src/lib/components/FigmaOverviewView.svelte → src/lib/components/OverviewView.svelte

  The orphan dial-era OverviewView was deleted above so the canonical name
  is free. FigmaOverviewView was an awkward "named after design source"
  name that aged badly the moment it shipped; the rename restores the
  intuitive name. Layout.svelte's import + mount updated; the
  figma-redesign-guardrails test assertion updated to look for
  'OverviewView' instead of 'FigmaOverviewView'.

Cross-cutting test updates (per spec's "grep tests/" protocol):

- tests/visual/overview-no-raw-url.spec.ts — Status and Live view sweeps
  removed (their .rail-row-label / .feed-name selectors no longer exist).
  Investigate-view sweep kept and updated to wait for .diagnose-title-name
  directly. Sentinel test updated to inject .diagnose-title-name (the only
  remaining primary-identifier selector).
- tests/visual/latency-scale.spec.ts — AC8 Status test (Dial +
  RacingStrip axis-label parity) removed; both components are deleted.
  The same invariant for the synthesis Overview surface will be
  re-established in PR 6 when NetworkTopology lands.
- tests/unit/figma-redesign-guardrails.test.ts — Layout assertion
  updated for the rename.

.stylelintignore updated: PR 5 deletion entries collapsed to a comment
record (the paths no longer exist); FigmaOverviewView entry replaced
with OverviewView (the rename target inherits the violations to be
cleaned in PR 6).

Verification:
- typecheck (tsc + svelte-check + functions): pass.
- lint (ESLint): pass.
- lint:css (stylelint): pass.
- test: 1115 / 1115 (was 1190; -75 from deleted test files).
- build: pass (~336 KB JS, was ~336 KB — no functional reduction since the
  dead code was already tree-shaken by Vite's bundler; deletion is a
  source-tree cleanup).

Net diff: 15 files deleted, 4,148 lines net removed (5,218 deletions
across deleted files + tests, ~50 lines added back across the rename
metadata, the .stylelintignore restructure, and the visual-test updates).
